splits dynamic/comp services interfaces in director_v2 plugin adapts test mocks
