+++ services/web/server/src/simcore_service_webserver/projects/projects_handlers.py
@@ -15,7 +15,7 @@ from servicelib.json_serialization import json_dumps
 from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
 
 from .._meta import api_version_prefix as VTAG
-from ..director_v2_core import DirectorServiceError
+from ..director_v2_exceptions import DirectorServiceError
 from ..login.decorators import login_required
 from ..security_decorators import permission_required
 from . import projects_api
