+++ services/web/server/tests/unit/with_dbs/03/version_control/conftest.py
@@ -5,7 +5,7 @@
 import logging
 from copy import deepcopy
 from pathlib import Path
-from typing import Any, AsyncIterator, Awaitable, Callable, Dict
+from typing import Any, AsyncIterator, Awaitable, Callable
 from unittest import mock
 from uuid import UUID
 
@@ -29,7 +29,7 @@ from simcore_service_webserver.db_models import UserRole
 from simcore_service_webserver.log import setup_logging
 from tenacity import AsyncRetrying, stop_after_delay
 
-ProjectDict = Dict[str, Any]
+ProjectDict = dict[str, Any]
 
 # HELPERS
 
@@ -77,7 +77,7 @@ async def catalog_subsystem_mock(monkeypatch, fake_project) -> None:
 @pytest.fixture
 def app_cfg(
     default_app_cfg, unused_tcp_port_factory, catalog_subsystem_mock, monkeypatch
-) -> Dict[str, Any]:
+) -> dict[str, Any]:
     """App's configuration used for every test in this module
 
     NOTE: Overrides services/web/server/tests/unit/with_dbs/conftest.py::app_cfg to influence app setup
@@ -200,7 +200,7 @@ async def request_delete_project(
         "simcore_service_webserver.projects.projects_api.director_v2_api.delete_pipeline",
     )
     director_v2_api_stop_services: mock.AsyncMock = mocker.patch(
-        "simcore_service_webserver.projects.projects_api.director_v2_api.stop_services",
+        "simcore_service_webserver.projects.projects_api.director_v2_api.stop_dynamic_services",
     )
     fire_and_forget_call_to_storage: mock.Mock = mocker.patch(
         "simcore_service_webserver.projects._delete.delete_data_folders_of_project",
